m_useJob
