+++ Assets/VRM10_Samples/ClothSample/ClothViewer/TextFields.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 58a0b5125aeceaa41b333a57e049c99e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
